handler/light_client: fix module structure
