+++ modules/handler/src/light_client/mod.rs
@@ -1,20 +1,14 @@
 pub use errors::LightClientHandlerError;
 pub use init_client::init_client;
+pub use query::query_client;
 pub use router::dispatch;
 pub use update_client::update_client;
-
-pub use {
-    query_client::query_client, verify_channel::verify_channel, verify_client::verify_client,
-    verify_client_consensus::verify_client_consensus, verify_connection::verify_connection,
-};
+pub use verify_state::{verify_channel, verify_client, verify_client_consensus, verify_connection};
 
 mod errors;
 mod init_client;
-mod query_client;
+mod query;
 mod registry;
 mod router;
 mod update_client;
-mod verify_channel;
-mod verify_client;
-mod verify_client_consensus;
-mod verify_connection;
+mod verify_state;
